refactor: move extensions and tools to main source set
