+++ it/src/main/java/org/eclipse/jkube/integrationtests/cli/CliUtils.java
