+++ it/src/test/java/org/eclipse/jkube/integrationtests/windows/WindowsITCase.java
@@ -18,7 +18,7 @@ import okhttp3.Request;
 import okhttp3.Response;
 import org.apache.maven.shared.invoker.InvocationResult;
 import org.apache.maven.shared.invoker.MavenInvocationException;
-import org.eclipse.jkube.integrationtests.docker.RegistryExtension;
+import org.eclipse.jkube.integrationtests.jupiter.api.extension.RegistryExtension;
 import org.eclipse.jkube.integrationtests.maven.BaseMavenCase;
 import org.eclipse.jkube.integrationtests.maven.MavenInvocationResult;
 import org.junit.jupiter.api.DisplayName;
